Delete DaemonClient layer, replace with EventStreamClient + DaemonStatus (Phase 4)
DaemonClient was a 1,225-line intermediary (across 3 files) that sat between
HTTPTransport and consumers. All its responsibilities are now handled by
focused, smaller types:

- EventStreamClient: SSE via GatewayHTTPClient.stream(), subscribe() broadcast,
  conversation ID mapping, sendUserMessage() fire-and-forget
- DaemonStatus: @Published state (isConnected, daemonVersion, etc.), connection
  lifecycle (health checks via HTTPTransport, auto-wake), message pre-processing

Key changes:
- ~50 onX callback closures replaced with subscribe() loops in consumers
- send(UserMessageMessage) replaced with sendUserMessage() (fire-and-forget)
- SSE moved from HTTPTransport to EventStreamClient (backed by GatewayHTTPClient)
- HTTPTransport slimmed to health checks + host tool execution only
- DaemonClient/DaemonClientProtocol kept as typealiases for compatibility
