+++ clients/macos/vellum-assistantTests/ConversationManagerRenameTests.swift
@@ -14,16 +14,11 @@ final class ConversationManagerRenameTests: XCTestCase {
         daemonClient = DaemonClient()
         daemonClient.isConnected = true
         capturedMessages = []
-        daemonClient.sendOverride = { [weak self] msg in
-            guard let self else { return }
-            capturedMessages.append(msg)
-        }
         conversationManager = ConversationManager(daemonClient: daemonClient)
         conversationManager.createConversation()
     }
 
     override func tearDown() {
-        daemonClient?.sendOverride = nil
         conversationManager = nil
         capturedMessages = []
         daemonClient = nil
