+++ clients/macos/vellum-assistantTests/VoiceModeManagerExtendedTests.swift
@@ -19,7 +19,6 @@ final class VoiceModeManagerExtendedTests: XCTestCase {
         manager = VoiceModeManager(voiceService: mockVoiceService)
         daemonClient = DaemonClient()
         daemonClient.isConnected = true
-        daemonClient.sendOverride = { _ in }
         chatViewModel = ChatViewModel(daemonClient: daemonClient)
     }
