--- clients/shared/Network/DaemonClient.swift
@@ -1,499 +1,0 @@
-import Foundation
-import os
-
-private let log = Logger(subsystem: Bundle.main.bundleIdentifier ?? "com.vellum.vellum-assistant", category: "DaemonClient")
-
-/// Shared signpost log for network instrumentation (Points of Interest lane in Instruments).
-private let networkLog = OSLog(
-    subsystem: Bundle.main.bundleIdentifier ?? "com.vellum.vellum-assistant",
-    category: .pointsOfInterest
-)
-
-/// Protocol for daemon client communication, enabling dependency injection and testing.
-@MainActor
-public protocol DaemonClientProtocol {
-    var isConnected: Bool { get }
-    func subscribe() -> AsyncStream<ServerMessage>
-    func send<T: Encodable>(_ message: T) throws
-    func connect() async throws
-    func disconnect()
-    func startSSE()
-    func stopSSE()
-}
-
-/// Platform-agnostic client for communicating with the Vellum daemon via HTTP + SSE.
-///
-/// This is a long-lived singleton. Consumers call `subscribe()` to get an independent message
-/// stream, enabling multiple consumers (HostCuExecutor, AmbientAgent) to each receive all
-/// messages and filter for the ones relevant to them.
-///
-/// - Important: New HTTP API calls should **not** be added here. Use `GatewayHTTPClient`
-///   instead, injected via a focused protocol (e.g. `ConversationClientProtocol`).
-///   Existing methods are being incrementally migrated to standalone clients backed by
-///   `GatewayHTTPClient`. See `clients/ARCHITECTURE.md` for details.
-@MainActor
-public final class DaemonClient: ObservableObject, DaemonClientProtocol {
-
-    // MARK: - Published State
-
-    @Published public var isConnected: Bool = false
-    public var isConnecting: Bool = false
-
-    /// The runtime HTTP server port, populated via `daemon_status` on connect.
-    /// `nil` means the HTTP server is not running.
-    @Published public var httpPort: Int?
-
-    /// Platform identifier for automatic 401 re-bootstrap (e.g. "macos", "ios").
-    /// Set by the app delegate after creating the client.
-    public var recoveryPlatform: String?
-
-    /// Device identifier for automatic 401 re-bootstrap.
-    /// Set by the app delegate after creating the client.
-    public var recoveryDeviceId: String?
-
-    /// Returns a closure that resolves the current HTTP port at call time.
-    /// Use this instead of reading `httpPort` directly when the value must
-    /// reflect the latest daemon state (e.g. after a daemon restart). The
-    /// closure captures `self` weakly to avoid retain cycles.
-    public var httpPortResolver: () -> Int? {
-        { [weak self] in self?.httpPort }
-    }
-
-    /// The daemon version string, populated via `daemon_status` on connect.
-    @Published public internal(set) var daemonVersion: String?
-
-    /// Whether the connected daemon's major.minor version differs from this client's version.
-    /// Set automatically when `daemon_status` is received. Does not block the connection.
-    @Published public internal(set) var versionMismatch: Bool = false
-
-    /// Whether a planned service group update is in progress.
-    /// Set when a `service_group_update_starting` event is received,
-    /// cleared when reconnected and `daemon_status` confirms the new version.
-    @Published public internal(set) var isUpdateInProgress: Bool = false
-
-    /// The version being upgraded to, if an update is in progress.
-    @Published public internal(set) var updateTargetVersion: String?
-
-    /// Deadline after which `isUpdateInProgress` is considered stale.
-    /// Computed from `expectedDowntimeSeconds` (with a 2x safety buffer)
-    /// when a `service_group_update_starting` event arrives. If the update
-    /// hasn't completed by this time, auto-wake suppression expires so the
-    /// client can recover from a crashed update.
-    var updateExpiresAt: Date?
-
-    /// Signing key fingerprint from the connected daemon, populated via `daemon_status`.
-    /// Used to detect instance switches — if this changes, the stored actor token is stale.
-    @Published public internal(set) var keyFingerprint: String?
-
-    /// Latest memory health payload from daemon `memory_status` events.
-    @Published public var latestMemoryStatus: MemoryStatusMessage?
-
-    /// Whether a TrustRulesView sheet is currently open from any settings surface.
-    /// Used to prevent multiple trust rules sheets from racing on the shared callback.
-    @Published public var isTrustRulesSheetOpen: Bool = false
-
-    // MARK: - Surface Event Callbacks
-
-    /// Called when the daemon sends a `ui_surface_show` message.
-    /// Set by the app layer to forward to SurfaceManager without coupling DaemonClient to it.
-    public var onSurfaceShow: ((UiSurfaceShowMessage) -> Void)?
-
-    /// Called when the daemon sends a `ui_surface_update` message.
-    public var onSurfaceUpdate: ((UiSurfaceUpdateMessage) -> Void)?
-
-    /// Called when the daemon sends a `ui_surface_dismiss` message.
-    public var onSurfaceDismiss: ((UiSurfaceDismissMessage) -> Void)?
-
-    /// Called when the daemon sends a `ui_surface_complete` message.
-    public var onSurfaceComplete: ((UiSurfaceCompleteMessage) -> Void)?
-
-    /// Called when the daemon sends a `document_editor_show` message.
-    public var onDocumentEditorShow: ((DocumentEditorShowMessage) -> Void)?
-
-    /// Called when the daemon sends a `document_editor_update` message.
-    public var onDocumentEditorUpdate: ((DocumentEditorUpdateMessage) -> Void)?
-
-    /// Called when the daemon sends a `document_save_response` message.
-    public var onDocumentSaveResponse: ((DocumentSaveResponseMessage) -> Void)?
-
-    /// Called when the daemon sends a `document_load_response` message.
-    public var onDocumentLoadResponse: ((DocumentLoadResponseMessage) -> Void)?
-
-    /// Called when the daemon sends an `app_files_changed` broadcast.
-    public var onAppFilesChanged: ((String) -> Void)?
-
-    /// Called when the daemon sends an `app_data_response` message.
-    public var onAppDataResponse: ((AppDataResponseMessage) -> Void)?
-
-    /// Called when the daemon sends a `message_queued` message.
-    public var onMessageQueued: ((MessageQueuedMessage) -> Void)?
-
-    /// Called when the daemon sends a `message_dequeued` message.
-    public var onMessageDequeued: ((MessageDequeuedMessage) -> Void)?
-
-    /// Called when the daemon sends a `message_queued_deleted` message.
-    public var onMessageQueuedDeleted: ((MessageQueuedDeletedMessage) -> Void)?
-
-    /// Called when the daemon sends a `generation_handoff` message.
-    public var onGenerationHandoff: ((GenerationHandoffMessage) -> Void)?
-
-    /// Called when the daemon sends a `confirmation_request` message for tool permission approval.
-    public var onConfirmationRequest: ((ConfirmationRequestMessage) -> Void)?
-
-    /// Called when the daemon sends a `confirmation_state_changed` message with authoritative state transitions.
-    public var onConfirmationStateChanged: ((ConfirmationStateChangedMessage) -> Void)?
-
-    /// Called when the daemon sends an `assistant_activity_state` message for thinking/streaming lifecycle.
-    public var onAssistantActivityState: ((AssistantActivityStateMessage) -> Void)?
-
-    /// Called when the daemon sends a `secret_request` message for secure credential input.
-    public var onSecretRequest: ((SecretRequestMessage) -> Void)?
-
-    /// Called when the daemon sends a `host_bash_request` message for proxy command execution.
-    public var onHostBashRequest: ((HostBashRequest) -> Void)?
-
-    /// Called when the daemon sends a `host_file_request` message for proxy file operations.
-    public var onHostFileRequest: ((HostFileRequest) -> Void)?
-
-    /// Called when the daemon sends a `host_cu_request` message for proxy CU action execution.
-    public var onHostCuRequest: ((HostCuRequest) -> Void)?
-
-    /// Called when the daemon sends a `dictation_response` message.
-    public var onDictationResponse: ((DictationResponseMessage) -> Void)?
-
-    /// Called when the daemon emits a generic `notification_intent` payload.
-    public var onNotificationIntent: ((NotificationIntentMessage) -> Void)?
-
-    /// Called when a notification delivery creates a new vellum conversation.
-    public var onNotificationConversationCreated: ((NotificationConversationCreated) -> Void)?
-
-    /// Called when the daemon broadcasts that a service group update is starting.
-    public var onServiceGroupUpdateStarting: ((ServiceGroupUpdateStartingMessage) -> Void)?
-
-    /// Called when the daemon broadcasts that a service group update has completed.
-    public var onServiceGroupUpdateComplete: ((ServiceGroupUpdateCompleteMessage) -> Void)?
-
-    /// Called when the server-assigned conversation ID differs from the
-    /// client-local ID. Parameters: (localId, serverId).
-    public var onConversationIdResolved: ((_ localId: String, _ serverId: String) -> Void)?
-
-    /// Called when the daemon sends a `skills_state_changed` push event.
-    public var onSkillStateChanged: ((SkillStateChangedMessage) -> Void)?
-
-    /// Called when the daemon sends a `trace_event` message.
-    public var onTraceEvent: ((TraceEventMessage) -> Void)?
-
-    /// Called when the daemon sends a `usage_update` message.
-    public var onUsageUpdate: ((UsageUpdate) -> Void)?
-
-    /// Called when the daemon sends a `conversation_list_response` message.
-    public var onConversationListResponse: ((ConversationListResponseMessage) -> Void)?
-
-    /// Called when the daemon sends a `conversation_title_updated` message.
-    public var onConversationTitleUpdated: ((ConversationTitleUpdatedMessage) -> Void)?
-
-    /// Called when the daemon sends a `history_response` message.
-    public var onHistoryResponse: ((HistoryResponse) -> Void)?
-
-    /// Called when the daemon sends a `slack_webhook_config_response` message.
-    public var onSlackWebhookConfigResponse: ((SlackWebhookConfigResponseMessage) -> Void)?
-
-    /// Called when the daemon sends an `ingress_config_response` message.
-    public var onIngressConfigResponse: ((IngressConfigResponseMessage) -> Void)?
-
-    /// Called when the daemon sends a `platform_config_response` message.
-    public var onPlatformConfigResponse: ((PlatformConfigResponseMessage) -> Void)?
-
-    /// Called when the daemon sends a `telegram_config_response` message.
-    public var onTelegramConfigResponse: ((TelegramConfigResponseMessage) -> Void)?
-
-
-    /// The currently active model ID, populated via `model_info` responses.
-    @Published public var currentModel: String?
-
-    /// The latest full model info response from the daemon stream.
-    @Published public var latestModelInfo: ModelInfoMessage?
-
-    /// Called when the daemon sends an `open_url` message.
-    public var onOpenUrl: ((OpenUrlMessage) -> Void)?
-
-    /// Called when the daemon sends a `navigate_settings` message.
-    public var onNavigateSettings: ((NavigateSettings) -> Void)?
-
-    /// Called when the daemon sends a `ui_layout_config` message.
-    public var onLayoutConfig: ((UiLayoutConfigMessage) -> Void)?
-
-    /// Called when the daemon sends a `diagnostics_export_response` message.
-    public var onDiagnosticsExportResponse: ((DiagnosticsExportResponseMessage) -> Void)?
-
-    /// Called when the daemon sends an `env_vars_response` message (debug builds only).
-    public var onEnvVarsResponse: ((EnvVarsResponseMessage) -> Void)?
-
-
-    /// Called when the daemon sends a `work_item_status_changed` broadcast.
-    public var onWorkItemStatusChanged: ((WorkItemStatusChanged) -> Void)?
-
-    /// Called when the daemon sends a `tasks_changed` broadcast.
-    public var onTasksChanged: ((TasksChanged) -> Void)?
-
-
-
-
-
-
-    /// Called when the daemon sends a generic `error` message (e.g. when a handler fails).
-    public var onError: ((ErrorMessage) -> Void)?
-
-    /// Called when a task run creates a conversation so the client can show it as a visible chat conversation.
-    public var onTaskRunConversationCreated: ((TaskRunConversationCreated) -> Void)?
-
-    /// Called when a schedule creates a conversation so the client can show it as a visible chat conversation.
-    public var onScheduleConversationCreated: ((ScheduleConversationCreated) -> Void)?
-
-    /// Called when the daemon requests pairing approval from macOS.
-    public var onPairingApprovalRequest: ((PairingApprovalRequestMessage) -> Void)?
-
-    /// Called when a subagent is spawned.
-    public var onSubagentSpawned: ((SubagentSpawned) -> Void)?
-
-    /// Called when a subagent's status changes (running, completed, failed, aborted).
-    public var onSubagentStatusChanged: ((SubagentStatusChanged) -> Void)?
-
-
-    /// Called when the daemon sends a `recording_pause` message.
-    public var onRecordingPause: ((RecordingPause) -> Void)?
-
-    /// Called when the daemon sends a `recording_resume` message.
-    public var onRecordingResume: ((RecordingResume) -> Void)?
-
-    /// Called when the daemon sends a `recording_start` message.
-    public var onRecordingStart: ((RecordingStart) -> Void)?
-
-    /// Called when the daemon sends a `recording_stop` message.
-    public var onRecordingStop: ((RecordingStop) -> Void)?
-
-    /// Called when the daemon sends a `client_settings_update` message.
-    public var onClientSettingsUpdate: ((ClientSettingsUpdate) -> Void)?
-
-    /// Called when the daemon broadcasts an `identity_changed` event (IDENTITY.md changed on disk).
-    public var onIdentityChanged: ((IdentityChanged) -> Void)?
-
-    /// Called when the daemon sends an `avatar_updated` message after regenerating the avatar.
-    public var onAvatarUpdated: ((AvatarUpdated) -> Void)?
-
-    /// Called when the daemon sends a `generate_avatar_response` message.
-    public var onGenerateAvatarResponse: ((GenerateAvatarResponse) -> Void)?
-
-    /// Called when the daemon sends a `contacts_response` message.
-    public var onContactsResponse: ((ContactsResponseMessage) -> Void)?
-
-    /// Called when the daemon broadcasts a `contacts_changed` event (contact table mutated).
-    public var onContactsChanged: ((ContactsChanged) -> Void)?
-
-    // MARK: - Auto-Wake
-
-    /// Optional closure invoked when a connection attempt fails because the daemon process
-    /// is not alive. The macOS app sets this to call `vellumCli.wake(name:)` so the
-    /// daemon is automatically restarted before retrying the connection.
-    /// Set by the app layer — `DaemonClient` never imports platform-specific types.
-    public var wakeHandler: (@MainActor @Sendable () async throws -> Void)?
-
-    #if os(macOS)
-    /// Timestamp of the last auto-wake attempt from the health-check disconnect path.
-    /// Used to prevent crash loops: if the daemon dies again within the cooldown window
-    /// after a wake, we stop retrying. Expires naturally after the cooldown period.
-    var lastAutoWakeAttempt: Date?
-
-    /// The in-flight auto-wake task, stored so it can be cancelled on intentional
-    /// disconnect or reconfigure to prevent reconnecting after teardown.
-    var autoWakeTask: Task<Void, Never>?
-    #endif
-
-    // MARK: - Broadcast Subscribers
-
-    /// Creates a new message stream for the caller. Each subscriber receives all messages
-    /// independently, enabling multiple consumers (HostCuExecutor, AmbientAgent) to
-    /// filter for messages relevant to them without competing for elements.
-    public func subscribe() -> AsyncStream<ServerMessage> {
-        let id = UUID()
-        let (stream, continuation) = AsyncStream<ServerMessage>.makeStream()
-        subscribers[id] = continuation
-        // onTermination fires on an arbitrary thread, but `subscribers` is
-        // MainActor-isolated. Dispatching via Task { @MainActor } is correct —
-        // the removal happens on the next MainActor tick, which is safe because
-        // a terminated continuation ignores further yields. The weak capture
-        // prevents a retain cycle if DaemonClient is deallocated first.
-        continuation.onTermination = { [weak self] _ in
-            Task { @MainActor [weak self] in
-                self?.subscribers.removeValue(forKey: id)
-            }
-        }
-        return stream
-    }
-
-    // MARK: - Internal State (accessed by extensions in DaemonConnection.swift and DaemonMessageRouter.swift)
-
-    var subscribers: [UUID: AsyncStream<ServerMessage>.Continuation] = [:]
-
-    var isAuthenticated = false
-
-    /// HTTP transport for communicating with the assistant.
-    public var httpTransport: HTTPTransport?
-
-    public private(set) var config: DaemonConfig
-
-    // MARK: - Init
-
-    public init(config: DaemonConfig = .default) {
-        self.config = config
-    }
-
-    // MARK: - Reconfigure
-
-    /// Reconfigure the daemon client's transport in place without replacing
-    /// the object identity. This preserves all callback closures and
-    /// subscriber references held by long-lived objects (ConversationManager,
-    /// ChatViewModel, RecordingManager, etc.) across assistant switches.
-    ///
-    /// The method disconnects the current transport, updates the config,
-    /// and resets connection-specific state. Callers must call `connect()`
-    /// after reconfiguring to establish the new connection.
-    public func reconfigure(config newConfig: DaemonConfig) {
-        #if os(macOS)
-        autoWakeTask?.cancel()
-        autoWakeTask = nil
-        #endif
-        disconnect()
-        self.config = newConfig
-        // Reset connection-specific state
-        isAuthenticated = false
-        httpPort = nil
-        daemonVersion = nil
-        versionMismatch = false
-        isUpdateInProgress = false
-        updateTargetVersion = nil
-        updateExpiresAt = nil
-        keyFingerprint = nil
-        latestMemoryStatus = nil
-        currentModel = nil
-        #if os(macOS)
-        lastAutoWakeAttempt = nil
-        #endif
-    }
-
-    /// Extract (major, minor) from a semver string like "1.2.3" or "v1.2.3".
-    private func parseMajorMinor(_ version: String) -> (Int, Int)? {
-        let cleaned = version.hasPrefix("v") ? String(version.dropFirst()) : version
-        let components = cleaned.split(separator: ".").compactMap { Int($0) }
-        guard components.count >= 2 else { return nil }
-        return (components[0], components[1])
-    }
-
-    /// Compare client and daemon major.minor versions. Logs a warning and sets
-    /// `versionMismatch` if they differ. Patch version differences are tolerated.
-    func checkVersionCompatibility(daemonVersion: String) {
-        guard let clientVersion = Bundle.main.infoDictionary?["CFBundleShortVersionString"] as? String else {
-            return
-        }
-        guard let (daemonMajor, daemonMinor) = parseMajorMinor(daemonVersion),
-              let (clientMajor, clientMinor) = parseMajorMinor(clientVersion) else {
-            return
-        }
-        let mismatch = daemonMajor != clientMajor || daemonMinor != clientMinor
-        if mismatch != versionMismatch {
-            versionMismatch = mismatch
-        }
-        if mismatch {
-            log.warning("Version mismatch: client \(clientVersion, privacy: .public) vs daemon \(daemonVersion, privacy: .public) — major.minor differs, features may not work correctly")
-        }
-    }
-
-    deinit {
-        // Swift 5.9+: deinit on @MainActor class is NOT guaranteed to run on main actor.
-        // Only call thread-safe cancellation methods here — Task.cancel() is safe from any thread.
-        #if os(macOS)
-        autoWakeTask?.cancel()
-        #endif
-        //
-        // We must finish subscriber continuations to prevent hanging `for await` loops.
-        // deinit guarantees exclusive access (no other strong references exist), so
-        // direct property access is safe without actor isolation dispatch.
-        let continuations = subscribers.values
-        for continuation in continuations {
-            continuation.finish()
-        }
-        // httpTransport is cleaned up via disconnectInternal() before dealloc;
-    }
-
-    // MARK: - Send
-
-    public enum SendError: Error, LocalizedError {
-        case notConnected
-
-        public var errorDescription: String? {
-            switch self {
-            case .notConnected:
-                return "Cannot send: not connected to assistant"
-            }
-        }
-    }
-
-    /// Legacy authentication errors — retained for compatibility with code
-    /// that catches `AuthError` (e.g. bootstrap retry coordinator).
-    public enum AuthError: Error, LocalizedError {
-        case missingToken
-        case timeout
-        case rejected(String?)
-
-        public var errorDescription: String? {
-            switch self {
-            case .missingToken:
-                return "Missing daemon session token"
-            case .timeout:
-                return "Daemon authentication timed out"
-            case .rejected(let message):
-                return message ?? "Daemon authentication rejected"
-            }
-        }
-    }
-
-    /// Closure that, when set, replaces the real send path.
-    /// Used in tests to avoid needing a live HTTP connection.
-    internal var sendOverride: ((Any) throws -> Void)?
-
-    /// Send a message to the daemon via HTTP transport.
-    /// Throws `SendError.notConnected` when the transport is unavailable.
-    public func send<T: Encodable>(_ message: T) throws {
-        let sendID = OSSignpostID(log: networkLog)
-        os_signpost(.begin, log: networkLog, name: "daemonHTTPSend", signpostID: sendID)
-
-        if let override = sendOverride {
-            os_signpost(.end, log: networkLog, name: "daemonHTTPSend", signpostID: sendID)
-            try override(message)
-            return
-        }
-
-        guard let httpTransport else {
-            os_signpost(.end, log: networkLog, name: "daemonHTTPSend", signpostID: sendID)
-            log.warning("Cannot send: not connected")
-            throw SendError.notConnected
-        }
-
-        guard httpTransport.isConnected else {
-            os_signpost(.end, log: networkLog, name: "daemonHTTPSend", signpostID: sendID)
-            throw SendError.notConnected
-        }
-
-        do {
-            try httpTransport.send(message)
-            os_signpost(.end, log: networkLog, name: "daemonHTTPSend", signpostID: sendID)
-        } catch {
-            os_signpost(.end, log: networkLog, name: "daemonHTTPSend", signpostID: sendID)
-            throw error
-        }
-    }
-
-    // MARK: - Conversations
-
-
-}
